% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainGG.R
\name{create_GG}
\alias{create_GG}
\title{Create genetic groups regarding countries}
\usage{
create_GG(
  psInputFile,
  psBreedsFile = NULL,
  psCountryFile = NULL,
  pminGroupSize = 250,
  pnumberOfYears = 6,
  pglobal_minyear = 1900,
  pglobal_maxyear = 2100,
  pdiffYear = 3,
  pshiftYear = 2,
  psOutputFile,
  pbLog = FALSE
)
}
\arguments{
\item{psInputFile}{file including number of missing parents per year and breed}

\item{psBreedsFile}{Read a breed grouping file}

\item{psCountryFile}{country grouping file}

\item{pminGroupSize}{minimum number of animals required per genetic group}

\item{pnumberOfYears}{minimum number of birth years required per genetic group}

\item{pglobal_minyear}{global minimum year}

\item{pglobal_maxyear}{globar maximum year}

\item{pdiffYear}{difference in birth years between parents of bulls and parents of dams}

\item{pshiftYear}{shift between genetic groups of sire and genetic groups of dams; this is required so that they do not contain exactly the same animals}

\item{psOutputFile}{genetic group csv-file}
}
\description{
Overall function to create genetic groups regarding countries
}
