% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGG_WithCountry.R
\name{define_GG}
\alias{define_GG}
\title{Define genetic group for breeds}
\usage{
define_GG(
  psInputFile,
  psBreedsFile = NULL,
  psCountryFile = NULL,
  pminGroupSize = 250,
  pnumberOfYears = 6,
  pdiffYear = 3,
  pglobal_minyear = 1900,
  pglobal_maxyear = 2100,
  pbLog = FALSE
)
}
\arguments{
\item{psInputFile}{data frame including number of missing parents per year and breed}

\item{psBreedsFile}{breed grouping file}

\item{psCountryFile}{country grouping file}

\item{pminGroupSize}{minimum number of animals required per genetic group}

\item{pnumberOfYears}{minimum number of birth years required per genetic group}

\item{pglobal_minyear}{global minimum year}

\item{pglobal_maxyear}{globar maximum year}

\item{ppdiffYear}{difference in birth years between parents of bulls and parents of dams}
}
\description{
Identifying breeds that have to few versus enough missing sires.
This criteria is regulated with the option pminGroupSize.
}
