% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptionsGG.R
\name{define_GG_dam}
\alias{define_GG_dam}
\title{Create genetic groups for dams}
\usage{
define_GG_dam(
  psInputFile,
  pshiftYear = 2,
  pglobal_minyear = 1900,
  pglobal_maxyear = 2100,
  pbLog = FALSE
)
}
\arguments{
\item{psInputFile}{List of genetic group}

\item{pshiftYear}{shift between genetic groups of sire and genetic groups of dams; this is required so that they do not contain exactly the same animals}

\item{pglobal_minyear}{global minimum year}

\item{pglobal_maxyear}{globar maximum year}
}
\description{
Create genetic groups for dams
}
