% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGG_WithoutCountry.R
\name{define_GG_withoutCountry}
\alias{define_GG_withoutCountry}
\title{Define genetic group for breeds}
\usage{
define_GG_withoutCountry(
  psInputFile,
  pminGroupSize = 250,
  pnumberOfYears = 6,
  pglobal_minyear = 1900,
  pglobal_maxyear = 2100,
  pbLog = FALSE
)
}
\arguments{
\item{psInputFile}{data frame including number of missing parents per year and breed}

\item{pminGroupSize}{minimum number of animals required per genetic group}

\item{pnumberOfYears}{minimum number of birth years required per genetic group}

\item{pglobal_minyear}{global minimum year}

\item{pglobal_maxyear}{globar maximum year}
}
\description{
Identifying breeds that have to few versus enough missing sires.
This criteria is regulated with the option pminGroupSize.
}
