% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputGG.R
\name{output_GG}
\alias{output_GG}
\title{Write output files with definition of genetic group}
\usage{
output_GG(psInputFile, psOutputFile, pbLog = FALSE)
}
\arguments{
\item{psInputFile}{List of genetic group}

\item{psOutputFile}{genetic group csv-file}
}
\description{
Write output files with definition of genetic group
}
