% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGG.R
\name{plot_statGenGroupOutFile}
\alias{plot_statGenGroupOutFile}
\title{Plot statistics on missing parents}
\usage{
plot_statGenGroupOutFile(psInputFile, pBreed, pbLog = FALSE)
}
\arguments{
\item{psInputFile}{file including number of missing parents per year and breed}

\item{pBreed}{breed to plot (breed code of Qualitas)}
}
\description{
Plot a table of statistics on missing parents
}
