% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGG.R
\name{read_statGenGroupOutFile}
\alias{read_statGenGroupOutFile}
\title{Reading statistics on missing parents}
\usage{
read_statGenGroupOutFile(psInputFile, pbLog = FALSE)
}
\arguments{
\item{psInputFile}{file including number of missing parents per year and breed}
}
\description{
Read a table of statistics on missing parents
}
