% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptionsGG.R
\name{substract_diffYear}
\alias{substract_diffYear}
\title{Subtract pdiffYear for genetic groups of parents of bulls}
\usage{
substract_diffYear(
  psInputFile,
  pdiffYear = 3,
  pglobal_minyear = 1900,
  pglobal_maxyear = 2100,
  pbLog = FALSE
)
}
\arguments{
\item{psInputFile}{List of genetic group}

\item{pglobal_minyear}{global minimum year}

\item{pglobal_maxyear}{globar maximum year}

\item{ppdiffYear}{difference in birth years between parents of bulls and parents of dams}
}
\description{
Distinction between parents of bulls and parents of cows.
Because the parent of sires reach the same genetic level earlier than parents of cows.
}
